==> 4_testing_best_models.ipynb <==
## Код для тестирования моделей

#### Импортирование необходимых библиотек

import numpy as np
import pandas as pd
from tqdm import tqdm
import warnings
from keras.models import load_model
from sentence_transformers import SentenceTransformer
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, classification_report

### Код для тестирования best_model
P.s. необходимо запустить данный блок если вы хотите протестировать наилучшую по качеству модель

warnings.filterwarnings("ignore")

df = pd.read_csv('test.csv')
df = df.rename(columns={'Description': 'text'})

embedding_model = SentenceTransformer("all-mpnet-base-v2")

def get_sentence_embedding(model, text):
    embeddings = model.encode([text])  
    return embeddings[0]

tqdm.pandas() 
df['embedddings'] = df['text'].progress_apply(lambda text: get_sentence_embedding(embedding_model, text))

loaded_model = load_model('best_model.h5')

X_test = np.array(df['embedddings'].to_list())
y_test = np.array(df['Label'])
y_pred_categorical = loaded_model.predict(X_test)
y_pred = y_pred_categorical.argmax(axis=1)
report = classification_report(y_test, y_pred)
accuracy = accuracy_score(y_test, y_pred)

print("Отчет по каждому классу:")
print(report)

accuracy = accuracy_score(y_test, y_pred)
precision = precision_score(y_test, y_pred, average='macro')
recall = recall_score(y_test, y_pred, average='macro')
f1 = f1_score(y_test, y_pred, average='macro')

print("Метрика         | Значение")
print("--------------------------")
print(f"Accuracy        | {accuracy:.4f}")
print(f"Precision macro | {precision:.4f}")
print(f"Recall macro    | {recall:.4f}")
print(f"F1-score macro  | {f1:.4f}")


### Код для тестирования fast_model
P.s. необходимо запустить данный блок если вы хотите протестировать оптимальную по качеству и скорости модель

warnings.filterwarnings("ignore")

df = pd.read_csv('test.csv')
df = df.rename(columns={'Description': 'text'})

embedding_model = SentenceTransformer("all-MiniLM-L6-v2")

def get_sentence_embedding(model, text):
    embeddings = model.encode([text])  
    return embeddings[0]

tqdm.pandas() 
df['embedddings'] = df['text'].progress_apply(lambda text: get_sentence_embedding(embedding_model, text))

loaded_model = load_model('fast_model.h5')

X_test = np.array(df['embedddings'].to_list())
y_test = np.array(df['Label'])
y_pred_categorical = loaded_model.predict(X_test)
y_pred = y_pred_categorical.argmax(axis=1)
report = classification_report(y_test, y_pred)
accuracy = accuracy_score(y_test, y_pred)

print("Отчет по каждому классу:")
print(report)

accuracy = accuracy_score(y_test, y_pred)
precision = precision_score(y_test, y_pred, average='macro')
recall = recall_score(y_test, y_pred, average='macro')
f1 = f1_score(y_test, y_pred, average='macro')

print("Метрика         | Значение")
print("--------------------------")
print(f"Accuracy        | {accuracy:.4f}")
print(f"Precision macro | {precision:.4f}")
print(f"Recall macro    | {recall:.4f}")
print(f"F1-score macro  | {f1:.4f}")
